experimental new trace and funky mode (-trace, -funky)
Trace mode will capture and store traces of new findings at runtime, as
part of initial/calibration fuzz phase.

Funky mode is made to better support fuzzing and diagnosis of
non-deterministic targets. New findings are tested multiple times and
only accepted if we get a quorom on the new coverage. Inputs with
confirmed non-determinism are stored for further diagnosis.